refactor(web): move character collection into features/collection/characters/
Reorganize the character collection feature to make space for the
upcoming weapon collection management (#44).

- Move useCollection → characters/useCharacterCollection
- Move useCollectionApi → characters/useCharacterCollectionApi
- Move useCollectionStore → characters/useCharacterCollectionStore
- Move CharacterCard and CharacterFilters from components/ into the
  characters feature directory
- Update all imports in CharactersPage
